+++ experiments/billable/src/app/pages/InvoiceDetail/functions.ts
@@ -25,35 +25,6 @@ export async function saveInvoice(id: string, invoice: Omit<Invoice, 'items' | '
   })
 }
 
-export async function createInvoice() {
-
-  // grab the supplier name
-  // and the contact information
-  // what if the user doesn't have any invoices?
-  // we will eventually include an invoice template... maybe I should just shove that in a seperate function for now?
-  let lastInvoice = await db.invoice.findFirst({
-    where: {
-      userId: '1',
-    },
-    orderBy: {
-      createdAt: 'desc',
-    }
-  })
-
-  const newInvoice = await db.invoice.create({
-    data: {
-      number: (Number(lastInvoice?.number || 0) + 1).toString(),
-      supplierName: lastInvoice?.supplierName,
-      supplierContact: lastInvoice?.supplierContact,
-      notesA: lastInvoice?.notesA,
-      notesB: lastInvoice?.notesB,
-      taxes: lastInvoice?.taxes,
-      userId: '1'
-    }
-  })
-
-  return newInvoice
-}
 
 export async function generatePdf(id: string) {
   return 'x'
