+++ experiments/billable/src/app/pages/InvoiceList/CreateInvoiceButton.tsx
@@ -1,7 +1,39 @@
 "use client";
 
 import { useTransition } from "react";
-import { createInvoice } from "./pages/invoiceDetail/functions";
+import { db } from "../../../db";
+
+
+export async function createInvoice() {
+
+  // grab the supplier name
+  // and the contact information
+  // what if the user doesn't have any invoices?
+  // we will eventually include an invoice template... maybe I should just shove that in a seperate function for now?
+  let lastInvoice = await db.invoice.findFirst({
+    where: {
+      userId: '1',
+    },
+    orderBy: {
+      createdAt: 'desc',
+    }
+  })
+
+  const newInvoice = await db.invoice.create({
+    data: {
+      number: (Number(lastInvoice?.number || 0) + 1).toString(),
+      supplierName: lastInvoice?.supplierName,
+      supplierContact: lastInvoice?.supplierContact,
+      notesA: lastInvoice?.notesA,
+      notesB: lastInvoice?.notesB,
+      taxes: lastInvoice?.taxes,
+      userId: '1'
+    }
+  })
+
+  return newInvoice
+}
+
 
 export function CreateInvoiceButton() {
   const [isPending, startTransition] = useTransition();
