+++ experiments/billable/src/app/services/invoices.ts
@@ -1,6 +1,5 @@
 "use server"
 import { db } from '../../db'
-import { calculateSubtotal, calculateTaxes } from '../shared/invoice'
 
 export type InvoiceItem = {
   description: string,
@@ -14,39 +13,7 @@ export type InvoiceTaxes = {
 }
 
 
-export async function getInvoiceListSummary() {
 
-  const invoices = await db.invoice.findMany({
-    select: {
-      id: true,
-      number: true,
-      date: true,
-      status: true,
-      customer: true,
-    },
-    where: {
-      userId: '1',
-    }
-  }) ?? []
-
-
-
-  return invoices.map((invoice) => {
-
-    const { id, date, number, customer, status } = invoice
-
-    // const subtotal = calculateSubtotal(invoice.items as InvoiceItem[])
-    // const taxes = calculateTaxes(subtotal, invoice.taxes as InvoiceTaxItem[])
-
-    return {
-      id,
-      date,
-      number,
-      customer: customer?.split('\n')[0] || '',
-      status,
-    }
-  })
-}
 
 // NOTE (peterp, 2025-01-13): The userID will be optional, since we should have that available "somewhere" in the context.
 export async function getInvoice(id: string) {
